+++ octodns_transip/__init__.py
@@ -5,13 +5,15 @@
 from collections import defaultdict, namedtuple
 from logging import getLogger
 
+from requests.utils import is_ipv4_address
 from transip import TransIP
 from transip.exceptions import TransIPHTTPError
-from transip.v6.objects import DnsEntry
+from transip.v6.objects import DnsEntry, Nameserver
+from urllib3.util.ssl_ import is_ipaddress
 
 from octodns.provider import ProviderException
 from octodns.provider.base import BaseProvider
-from octodns.record import Record
+from octodns.record import Record, Update
 
 # TODO: remove __VERSION__ with the next major version release
 __version__ = __VERSION__ = '0.0.1'
@@ -34,15 +36,41 @@ class TransipNewZoneException(TransipException):
 class TransipProvider(BaseProvider):
     SUPPORTS_GEO = False
     SUPPORTS_DYNAMIC = False
+    SUPPORTS_ROOT_NS = False
     SUPPORTS = set(
-        ('A', 'AAAA', 'CNAME', 'MX', 'NS', 'SRV', 'SPF', 'TXT', 'SSHFP', 'CAA')
+        (
+            'A',
+            'AAAA',
+            'CNAME',
+            'MX',
+            'NS',
+            'SRV',
+            'SPF',
+            'TXT',
+            'SSHFP',
+            'CAA',
+            'TLSA',
+            'NAPTR',
+            'ALIAS',
+            'DS',
+        )
     )
-    # unsupported by OctoDNS: 'TLSA'
     MIN_TTL = 120
     TIMEOUT = 15
     ROOT_RECORD = '@'
+    ROOT_NS_TTL = 3600  # Bogus value
 
-    def __init__(self, id, account, key=None, key_file=None, *args, **kwargs):
+    def __init__(
+        self,
+        id,
+        account,
+        key=None,
+        key_file=None,
+        enable_root_ns=False,
+        *args,
+        **kwargs,
+    ):
+        self.SUPPORTS_ROOT_NS = enable_root_ns
         self.log = getLogger('TransipProvider[{}]'.format(id))
         self.log.debug('__init__: id=%s, account=%s, token=***', id, account)
         super().__init__(id, *args, **kwargs)
@@ -72,6 +100,10 @@ class TransipProvider(BaseProvider):
         try:
             domain = self._client.domains.get(zone.name.strip('.'))
             records = domain.dns.list()
+            nameservers = (
+                domain.nameservers.list() if self.SUPPORTS_ROOT_NS else []
+            )
+
         except TransIPHTTPError as e:
             if e.response_code == 404 and target is False:
                 # Zone not found in account, and not a target so just
@@ -96,6 +128,26 @@ class TransipProvider(BaseProvider):
         self.log.debug(
             'populate: found %s records for zone %s', len(records), zone.name
         )
+
+        if nameservers:
+            values = []
+            for ns in nameservers:
+                if ns.hostname != '':
+                    values.append(ns.hostname + '.')
+                if ns.ipv4 != '':
+                    values.append(ns.ipv4 + '.')
+                if ns.ipv6 != '':
+                    values.append(ns.ipv6 + '.')
+
+            record = Record.new(
+                zone,
+                '',
+                {'type': 'NS', 'ttl': self.ROOT_NS_TTL, 'values': values},
+                source=self,
+                lenient=lenient,
+            )
+            zone.add_record(record, lenient=lenient)
+
         if records:
             values = defaultdict(lambda: defaultdict(list))
             for record in records:
@@ -116,11 +168,18 @@ class TransipProvider(BaseProvider):
                         lenient=lenient,
                     )
                     zone.add_record(record, lenient=lenient)
+
         self.log.info(
             'populate:   found %s records', len(zone.records) - before
         )
 
         return True
+
+    def _process_desired_zone(self, desired):
+
+        self.log.info("process_desired")
+
+        return desired
 
     def _apply(self, plan):
         desired = plan.desired
@@ -135,11 +194,59 @@ class TransipProvider(BaseProvider):
                 'Unhandled error: ({}) {}'.format(e.response_code, e.message)
             )
 
+        for change in changes:
+            record = change.data['new']
+
+            if change.data['name'] == '' and change.data['record_type'] == 'NS':
+                values = (
+                    record.get('values')
+                    if record.get('values')
+                    else [record.get('value')]
+                )
+
+                # Root nameservers has no TTL within Transip, so skip apply if values are the same
+                # TODO: Find a way to exclude TTL-only change from the plan
+                if change.__class__ == Update:
+                    existing = change.data['existing']
+                    existing_values = (
+                        existing.get('values')
+                        if existing.get('values')
+                        else [existing.get('value')]
+                    )
+                    if existing_values == values:
+                        self.log.debug(
+                            "No change in root nameserver values, skipping"
+                        )
+                        continue
+
+                nameservers = []
+                for value in values:
+                    nameservers.append(
+                        Nameserver(
+                            domain.nameservers, _attr_for_nameserver(value)
+                        )
+                    )
+                try:
+                    domain.nameservers.replace(nameservers)
+                except TransIPHTTPError as e:
+                    self.log.warning(
+                        '_apply: Set Nameservers returned one or more errors: {}'.format(
+                            e
+                        )
+                    )
+                    raise TransipException(
+                        'Unhandled error: ({}) {}'.format(
+                            e.response_code, e.message
+                        )
+                    )
+
         records = []
         for record in plan.desired.records:
             if record._type in self.SUPPORTS:
                 # Root records have '@' as name
                 name = record.name
+                if name == '' and record._type == 'NS':
+                    continue
                 if name == '':
                     name = self.ROOT_RECORD
 
@@ -160,7 +267,7 @@ class TransipProvider(BaseProvider):
 
 
 def _data_for(type_, records, current_zone):
-    if type_ == 'CNAME':
+    if type_ == 'CNAME' or type_ == 'ALIAS':
         return {
             'type': type_,
             'ttl': records[0].expire,
@@ -198,12 +305,51 @@ def _data_for(type_, records, current_zone):
     def format_txt(record):
         return record.content.replace(';', '\\;')
 
+    def format_tlsa(record):
+        (
+            certificate_usage,
+            selector,
+            matching_type,
+            certificate_association_data,
+        ) = record.content.split(' ', 4)
+        return {
+            'certificate_usage': certificate_usage,
+            'selector': selector,
+            'matching_type': matching_type,
+            'certificate_association_data': certificate_association_data,
+        }
+
+    def format_naptr(record):
+        order, preference, flags, service, regexp, replacement = (
+            record.content.split(' ', 6)
+        )
+        return {
+            'order': order,
+            'preference': preference,
+            'flags': flags,
+            'service': service,
+            'regexp': regexp,
+            'replacement': replacement,
+        }
+
+    def format_ds(record):
+        key_tag, algorithm, digest_type, digest = record.content.split(' ', 4)
+        return {
+            'key_tag': key_tag,
+            'algorithm': algorithm,
+            'digest_type': digest_type,
+            'digest': digest,
+        }
+
     value_formatter = {
         'MX': format_mx,
         'SRV': format_srv,
         'SSHFP': format_sshfp,
         'CAA': format_caa,
         'TXT': format_txt,
+        'TLSA': format_tlsa,
+        'NAPTR': format_naptr,
+        'DS': format_ds,
     }.get(type_, lambda r: r.content)
 
     return {
@@ -231,16 +377,56 @@ def _get_lowest_ttl(records):
 
 def _entries_for(name, record):
     values = record.values if hasattr(record, 'values') else [record.value]
+
+    def entry_mx(v):
+        return f'{v.preference} {v.exchange}'
+
+    def entry_srv(v):
+        return f'{v.priority} {v.weight} {v.port} {v.target}'
+
+    def entry_sshfp(v):
+        return f'{v.algorithm} {v.fingerprint_type} {v.fingerprint}'
+
+    def entry_caa(v):
+        return f'{v.flags} {v.tag} {v.value}'
+
+    def entry_txt(v):
+        return v.replace('\\;', ';')
+
+    def entry_tlsa(v):
+        return f'{v.certificate_usage} {v.selector} {v.matching_type} {v.certificate_association_data}'
+
+    def entry_naptr(v):
+        return f'{v.order} {v.preference} {v.flags} {v.service} {v.regexp} {v.replacement}'
+
+    def entry_ds(v):
+        return f'{v.key_tag} {v.algorithm} {v.digest_type} {v.digest}'
+
     formatter = {
-        'MX': lambda v: f'{v.preference} {v.exchange}',
-        'SRV': lambda v: f'{v.priority} {v.weight} {v.port} {v.target}',
-        'SSHFP': lambda v: (
-            f'{v.algorithm} {v.fingerprint_type} {v.fingerprint}'
-        ),
-        'CAA': lambda v: f'{v.flags} {v.tag} {v.value}',
-        'TXT': lambda v: v.replace('\\;', ';'),
+        'MX': entry_mx,
+        'SRV': entry_srv,
+        'SSHFP': entry_sshfp,
+        'CAA': entry_caa,
+        'TXT': entry_txt,
+        'TLSA': entry_tlsa,
+        'NAPTR': entry_naptr,
+        'DS': entry_ds,
     }.get(record._type, lambda r: r)
+
     return [
         DNSEntry(name, record.ttl, record._type, formatter(value))
         for value in values
     ]
+
+
+def _attr_for_nameserver(nameserver):
+    nameserver = nameserver.strip('.')
+    return {
+        'hostname': nameserver if not is_ipaddress(nameserver) else '',
+        'ipv4': nameserver if is_ipv4_address(nameserver) else '',
+        'ipv6': (
+            nameserver
+            if is_ipaddress(nameserver) and not is_ipv4_address(nameserver)
+            else ''
+        ),
+    }