+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## UNRELEASED
+
+#### Nothworthy Changes
+
+* Added support for more record types: ALIAS, DS, NAPTR, TLSA, TLSA
+* Added support for root nameservers (opt-in)
+
 ## v0.0.1 - 2022-01-14 - Moving
 
 #### Nothworthy Changes